refactor: align reminder file layout and naming
