+++ src/melanki.trippeltrumf.service/Features/Notifying/Reminder.cs
@@ -28,25 +28,3 @@ public enum ReminderEligibility
     NotDayBefore = 2,
     AlreadySent = 3
 }
-
-public sealed class ReminderStateStore
-{
-    private readonly object _lock = new();
-    private readonly HashSet<DateOnly> _sentReminderDates = [];
-
-    public bool HasSent(DateOnly eventDate)
-    {
-        lock (_lock)
-        {
-            return _sentReminderDates.Contains(eventDate);
-        }
-    }
-
-    public bool TryMarkSent(DateOnly eventDate)
-    {
-        lock (_lock)
-        {
-            return _sentReminderDates.Add(eventDate);
-        }
-    }
-}
